HighDPI Fix
When switching over to GLFW i realized that the window size and frame
buffer size are different. The window size is based of position on the
screen and the frame buffer size is based of pixels. So i need to grab
the frame buffer size and set the view port there.